+++ packages/client-ledger/RELEASES.md
@@ -177,3 +177,13 @@ Based on:
 - [typescript v1.0.0-alpha.17] packages/client-ledger
 ### Releases
 - [NPM v1.0.0-alpha.17] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.17 - packages/client-ledger
+
+## 2024-08-09 00:16:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.358.0 (2.390.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.18] packages/client-ledger
+### Releases
+- [NPM v1.0.0-alpha.18] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.18 - packages/client-ledger